Move MaxRes algorithm to a separate solver
