+++ ksmt-maxsat/src/main/kotlin/io/ksmt/maxsat/SoftConstraint.kt
@@ -1,4 +1,4 @@
-package io.ksmt.solver.maxsat
+package io.ksmt.maxsat
 
 import io.ksmt.expr.KExpr
 import io.ksmt.sort.KBoolSort
